+++ CHANGELOG.md
@@ -15,6 +15,9 @@
 - Added Google Maps styles for marker mode
 - Added marker showing for location search 
 
+# 2/22/2021
+- Added responsive UI to placeList and map markers respectively
+
 # 2/26/2021
 - Started editMarker branch and editing options
 
@@ -22,3 +25,4 @@
 - Removed drag clicks persisting on markerWindows
 - Fixed editMarker edits creating new markers instead of updating
 - Removed buggy 'edit' clicks caused from clicking edit in edit mode
+